[Feat] #45 - 레시피 컬렉션뷰 레이아웃 함수 재사용
