+++ Mukjipsa-iOS/Mukjipsa-iOS/Screen/TabBar/Recipe/Reuse/ReuseRecipeCollectionLayout.swift
@@ -1,0 +1,45 @@
+//
+//  RecipeCollectionViewLayout.swift
+//  Mukjipsa-iOS
+//
+//  Created by 황찬미 on 2022/11/18.
+//
+
+import UIKit
+
+extension UIViewController {
+    func createSectionLayout() -> UICollectionViewLayout {
+        return UICollectionViewCompositionalLayout { (sectionNumber, _) -> NSCollectionLayoutSection? in
+            switch sectionNumber {
+            case 0:
+                return self.createRecipeLayout()
+            default:
+                return self.createRecipeLayout()
+            }
+        }
+    }
+    func createRecipeLayout() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1.0),
+            heightDimension: .estimated(375))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 16, trailing: 0)
+        let groupSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1.0),
+            heightDimension: .absolute(375))
+        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitem: item, count: 1)
+        let section = NSCollectionLayoutSection(group: group)
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
+        
+        let headerSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1.0),
+            heightDimension: .absolute(152))
+        let header = NSCollectionLayoutBoundarySupplementaryItem(
+            layoutSize: headerSize,
+            elementKind: UICollectionView.elementKindSectionHeader,
+            alignment: .top)
+
+        section.boundarySupplementaryItems = [header]
+        return section
+    }
+}
